tasterealm-data-guard/requirements.txt
# Database
psycopg2-binary>=2.9.11
sqlalchemy==2.0.25

# FastAPI (for later)
fastapi==0.109.0
uvicorn==0.27.0

# Data handling
pandas>=2.3.3

# Dev tools
ipython==8.20.0
black==24.1.1
